# lib/api-zod/src/generated/types/verifyClaim.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyClaimLabel } from "./verifyClaimLabel";
import type { VerifySource } from "./verifySource";

export interface VerifyClaim {
  number: number;
  text: string;
  label: VerifyClaimLabel;
  /** One-sentence plain-English explanation of the label */
  rationale: string;
  /**
   * What is established (for speculative/contested claims)
   * @nullable
   */
  established?: string | null;
  /**
   * What is not established (for speculative/contested claims)
   * @nullable
   */
  not_established?: string | null;
  sources: VerifySource[];
}