# lib/api-zod/src/generated/types/verifyJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyClaim } from "./verifyClaim";
import type { VerifyJobStatus } from "./verifyJobStatus";

export interface VerifyJob {
  job_id: string;
  status: VerifyJobStatus;
  /** 0-100 coarse progress */
  progress: number;
  /** Plain-English label for the current step */
  step_label: string;
  /**
   * Plain-English error message when status is failed
   * @nullable
   */
  error_message?: string | null;
  /** @nullable */
  video_title?: string | null;
  /** @nullable */
  video_thumbnail?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  duration_seconds?: number | null;
  /**
   * One-sentence plain-English summary of the video's subject
   * @nullable
   */
  summary?: string | null;
  /** @nullable */
  claims?: VerifyClaim[] | null;
  /** @nullable */
  share_slug?: string | null;
  /** @nullable */
  created_at?: Date | null;
}