# lib/api-zod/src/generated/types/verifyUrlInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyUrlInput {
  /**
   * Public video URL (YouTube, Facebook, Rumble, etc.)
   * @minLength 1
   * @maxLength 2000
   */
  url: string;
}